move user input plugin to common scope, game launchable using keyboard, mouse and touch, remove fullscreen breaking mobile browser
