+++ src/common/plugins/mod.rs
@@ -1,0 +1,1 @@
+pub mod user_input_plugin;
