+++ src/game/plugins/mod.rs
@@ -1,3 +1,2 @@
 pub mod game_scene_plugin;
 pub mod player_plugin;
-pub mod user_input_plugin;
